pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flow-factory"
version = "0.1.0"
description = "Unified RL Fine-tuning Framework for Diffusion/Flow-Matching Models"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Flow-Factory Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# Core Dependencies
dependencies = [
    # --- Deep Learning Core ---
    "torch>=2.6.0",
    "torchvision>=0.19.0",
    "torchaudio>=2.4.0",
    
    # --- Hugging Face Ecosystem ---
    "transformers>=4.44.0",
    "diffusers>=0.30.0",
    "accelerate>=0.33.0",
    "peft>=0.12.0",
    "datasets>=2.20.0",
    "tokenizers>=0.19.0",
    "huggingface-hub>=0.24.0",

    # --- Optimization & Training ---
    "deepspeed>=0.14.0",      # For ZeRO stage optimization
    "bitsandbytes>=0.43.0",   # For 8-bit/4-bit quantization
    "einops>=0.8.0",
    
    # --- Critical Fixes for Flux/T5 ---
    "protobuf>=6.33.2",       # REQUIRED for T5 serialization
    "sentencepiece>=0.2.0",   # REQUIRED for T5 tokenizer

    # --- Data Processing ---
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pillow>=10.4.0",
    "scipy>=1.13.0",
    "scikit-learn>=1.5.0",
    "scikit-image>=0.24.0",
    "albumentations>=1.4.0",
    "opencv-python>=4.10.0",

    # --- Logging & Utils ---
    "tqdm",
    "wandb",
    "swanlab",
    "pydantic>=2.8.0",
    "pyyaml>=6.0",
    "absl-py",
    "ml_collections",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "ipython",
]
# For NVIDIA specific optimized kernels (optional as they can be tricky to install)
nvidia = [
    "xformers>=0.0.27",
    "nvidia-ml-py",
]

[project.scripts]
flow-factory-train = "flow_factory.cli:train_cli"
ff-train = "flow_factory.cli:train_cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100